tests(district heating): add new unittests
