+++ tests/test_preprocessing_district_heating.py
@@ -195,82 +195,6 @@ def test_clear_thermal_net(dataframe_name):
     # Verifying that the specified dataframe is now an empty dataframe
     assert result.components[dataframe_name].empty
     
-
-def test_create_fork(test_forks_dataframe, thermal_net):
-    """
-        This test checks whether the row created by the create fork
-        method has been correctly parameterized in the DataFrame of the
-        thermal network forks.
-    """
-    # Creating an instance of the ThermalNetwork
-    # Assuming create_fork is a method of the district_heating module
-    # and it adds a new fork to the thermal network
-    # with the specified parameters (point, label, bus)
-    point = ["x", 10, 10, "x", 0.5, "test"]
-    label = 1
-    bus = "testbus"
-    
-    # Calling the create_fork method to add a new fork to the thermal
-    # network
-    thermal_net = district_heating.create_fork(
-        point=point,
-        label=label,
-        thermal_net=thermal_net,
-        bus=bus)
-
-    # Comparing the resulting 'forks' component DataFrame with the
-    # expected test dataframe
-    pandas.testing.assert_frame_equal(
-        left=thermal_net.components["forks"],
-        right=test_forks_dataframe
-    )
-
-
-@pytest.fixture
-def test_pipes_dataframe():
-    """
-        The pandas dataframe created here provides the benchmark for
-        test_append_pipe.
-    """
-    return pandas.DataFrame.from_dict(
-        {"id": ["pipe-1"],
-         "from_node": ["forks-1"],
-         "to_node": ["forks-2"],
-         "length": [30.0],
-         "component_type": ["Pipe"],
-         "street": ["test-street"]}
-    )
-
-
-def test_append_pipe(test_pipes_dataframe, thermal_net):
-    """
-        This test checks whether the row created by the append pipe
-        method has been correctly parameterized in the DataFrame of the
-        thermal network pipes.
-    """
-    # Creating an instance of the ThermalNetwork
-    # Assuming append_pipe is a method of the district_heating module
-    # and it adds a new pipe to the thermal network
-    # with the specified parameters (nodes, length, street)
-    nodes = ["forks-1", "forks-2"]
-    length = 30.0
-    street = "test-street"
-
-    # Calling the append_pipe method to add a new pipe to the thermal
-    # network
-    thermal_net = district_heating.append_pipe(
-        nodes=nodes,
-        length=length,
-        street=street,
-        thermal_net=thermal_net)
-
-    # Comparing the resulting 'pipes' component DataFrame with the
-    # expected test dataframe
-    pandas.testing.assert_frame_equal(
-        left=thermal_net.components["pipes"],
-        right=test_pipes_dataframe
-    )
-    
     
 class MockOemofInvestOptimizationModel:
     def __init__(self, nodes):
@@ -325,102 +249,7 @@ def test_remove_sinks_collect_buses(mock_oemof_model):
     assert "other_node" in [node.label for node in updated_model.nodes]
     
     
-def test_create_connection_points():
-    """
-        TODO
-    """
-    pass
 
-
-@pytest.fixture
-def street_sec():
-    return pandas.DataFrame({
-        'active': [1, 1],
-        'lat. 1st intersection': [1.0, 2.0],
-        'lon. 1st intersection': [3.0, 4.0],
-        'lat. 2nd intersection': [5.0, 6.0],
-        'lon. 2nd intersection': [7.0, 8.0]
-    })
-
-
-def test_create_intersection_forks(street_sec, thermal_net):
-    """
-        Test the create_intersection_forks function.
-
-        This test function checks the behavior of the
-        create_intersection_forks method by providing it with specific
-        street_sec and thermal_net inputs and then asserting the
-        expected behavior and properties of the resulting
-        ThermalNetwork instance.
-
-        :param: street_sec: Test DataFrame containing street \
-          sections with beginning and ending points.
-        :type street_sec: pandas.DataFrame
-        :param thermal_net: Test instance of the DHNx ThermalNetwork \
-          used as a basis for creating the components.
-        :type street_sec: dhnx.network.ThermalNetwork
-        """
-    # Call the method with the test data
-    result_thermal_net = district_heating.create_intersection_forks(
-        street_sec=street_sec, thermal_net=thermal_net)
-    
-    # Assert the result is an instance of ThermalNetwork
-    assert isinstance(result_thermal_net, dhnx.network.ThermalNetwork)
-    
-    # Check if the forks component is updated
-    assert len(result_thermal_net.components["forks"]) == 4
-    
-    # Check if the properties of the forks are as expected
-    forks = result_thermal_net.components["forks"]
-    assert sorted(forks["lat"].to_list()) == [1.0, 2.0, 5.0, 6.0]
-    assert sorted(forks["lon"].to_list()) == [3.0, 4.0, 7.0, 8.0]
     
     
-class MockThermalNetwork2:
-    def __init__(self, forks, pipes):
-        self.components = {"forks": forks,
-                           "pipes": pipes}
 
-
-@pytest.fixture
-def sample_points_data():
-    return MockThermalNetwork2(
-        forks=pandas.DataFrame(data={
-            "lat": [5.0, 10.0, 20.0, 12.0, 22.0],
-            "lon": [2.0, 15.0, 25.0, 18.0, 28.0],
-            "id": ["consumers-1", "forks-1", "forks-2", "forks-3", "forks-4"],
-            "t": ["", 0, 0.25, 0.5, 0.75]
-        }),
-        pipes=pandas.DataFrame(data={}),
-    )
-
-
-@pytest.fixture
-def sample_streets_data():
-    data = {
-        "label": ["street1", "street2"],
-        "active": [1, 1],
-        "lat. 1st intersection": [10.0, 20.0],
-        "lon. 1st intersection": [15.0, 25.0],
-        "lat. 2nd intersection": [12.0, 22.0],
-        "lon. 2nd intersection": [18.0, 28.0],
-    }
-    return pandas.DataFrame(data)
-
-
-def test_create_supply_line(sample_streets_data, sample_points_data):
-    # Call the method with sample data
-    result_thermal_net = district_heating.create_supply_line(
-        streets=sample_streets_data,
-        thermal_net=sample_points_data)
-
-    pipes = result_thermal_net.components["pipes"]
-    # Check if pipes where added to the thermal network pipes dataframe
-    assert len(pipes) > 0
-    
-    # Check if forks-1 -- a from node -- is listed in from node column
-    assert "forks-1" in pipes["from_node"].to_list()
-    # Check if length of pipe 1 is listed in length column
-    assert 396208.1053375902 in pipes["length"].to_list()
-    # Check if forks-3 -- a to node -- is listed in to node column
-    assert "forks-3" in pipes["to_node"].to_list()
